Refactor to use Ember Service
