+++ app/assets/javascripts/admin/models/theme.js
@@ -7,7 +7,6 @@ import discourseComputed from "discourse-common/utils/decorators";
 import { popupAjaxError } from "discourse/lib/ajax-error";
 import { ajax } from "discourse/lib/ajax";
 import { escapeExpression } from "discourse/lib/utilities";
-import highlightSyntax from "discourse/lib/highlight-syntax";
 import { url } from "discourse/lib/computed";
 
 const THEME_UPLOAD_VAR = 2;
@@ -321,7 +320,6 @@ const Theme = RestModel.extend({
             }
           }
         );
-        highlightSyntax(document.querySelector(".bootbox.modal"));
       } else {
         return this.save({ remote_update: true }).then(() =>
           this.set("changed", false)
