+++ test/javascripts/components/highlighted-code-test.js
@@ -1,5 +1,5 @@
 import componentTest from "helpers/component-test";
-import { waitForHighlighting } from "discourse/lib/highlight-syntax";
+import { waitForHighlighting } from "discourse/services/syntax-highlighter";
 
 const LONG_CODE_BLOCK = "puts a\n".repeat(15000);
 
@@ -18,20 +18,27 @@ componentTest("highlighting code", {
     this.set("code", "def test; end");
 
     await waitForHighlighting();
-
     assert.equal(
       find("code.ruby.hljs .hljs-function .hljs-keyword")
         .text()
         .trim(),
       "def"
     );
+  }
+});
+
+componentTest("highlighting code limit", {
+  template: "{{highlighted-code lang='ruby' code=code}}",
+
+  beforeEach() {
+    Discourse.HighlightJSPath =
+      "/assets/highlightjs/highlight-test-bundle.min.js";
+    Discourse.HighlightJSWorkerURL = "/assets/highlightjs-worker.js";
   },
 
-  
   async test(assert) {
     this.set("code", LONG_CODE_BLOCK);
     await waitForHighlighting();
-
     assert.equal(
       find("code")
         .text()
@@ -39,5 +46,4 @@ componentTest("highlighting code", {
       LONG_CODE_BLOCK.trim()
     );
   }
-
 });
